fix(analysis): address PR review — deadlock risk, test isolation, stale comments
1. [ARCHITECTURAL] Snapshot declarations under RLock then release
   before invoking user callback in HarvestDeclarations to prevent
   deadlock if callback calls other indexer methods.

2. [TECHNICAL DEBT] Move fixtureIndexer and IndexSnapshot into
   _test.go files so test infrastructure is excluded from the
   production binary.

3. [TECHNICAL DEBT] Fix stale comment referencing deleted
   harvestObjectSymbols — now points to collectDeclarations.
