--- internal/tools/analysis/index_fixture.go
@@ -1,142 +1,0 @@
-// Copyright (c) 2026 gosharplite@gmail.com
-// SPDX-License-Identifier: MIT
-
-package analysis
-
-import (
-	"context"
-	"sync"
-
-	"golang.org/x/tools/go/packages"
-)
-
-// fixtureIndexer implements symbolIndex from a pre-built IndexSnapshot.
-// It bypasses packages.Load entirely — all data is served from the snapshot.
-// Used in tests to avoid the expensive go/packages type-checking step.
-type fixtureIndexer struct {
-	mu       sync.RWMutex
-	snapshot *IndexSnapshot
-
-	// Pre-computed lookup maps for O(1) access.
-	declsByID map[string]*symMeta
-	pkgPaths  map[string]bool
-}
-
-// Compile-time interface check.
-var _ SymbolIndex = (*fixtureIndexer)(nil)
-
-// newFixtureIndexer creates a fixtureIndexer from an IndexSnapshot.
-// It pre-computes lookup maps for O(1) access during analysis.
-func newFixtureIndexer(s *IndexSnapshot) *fixtureIndexer {
-	fi := &fixtureIndexer{
-		snapshot:  s,
-		declsByID: make(map[string]*symMeta, len(s.Declarations)),
-		pkgPaths:  make(map[string]bool, len(s.Declarations)),
-	}
-	for _, decl := range s.Declarations {
-		fi.declsByID[decl.id] = decl
-		fi.pkgPaths[decl.pkgPath] = true
-	}
-	return fi
-}
-
-func (fi *fixtureIndexer) Refresh(ctx context.Context, hb chan<- struct{}) error {
-	return nil // Fixture is immutable; no-op.
-}
-
-func (fi *fixtureIndexer) Lookup(ctx context.Context, symbol string, hb chan<- struct{}) ([]location, error) {
-	return nil, nil
-}
-
-func (fi *fixtureIndexer) FindImplementors(ctx context.Context, interfaceName string, hb chan<- struct{}) ([]typeName, error) {
-	return nil, nil
-}
-
-func (fi *fixtureIndexer) SearchSymbols(ctx context.Context, path string, query string, exportedOnly bool, hb chan<- struct{}) ([]symbolLocation, error) {
-	return nil, nil
-}
-
-func (fi *fixtureIndexer) GetUsages(ctx context.Context, symbol string, path string, hb chan<- struct{}) ([]location, error) {
-	fi.mu.RLock()
-	defer fi.mu.RUnlock()
-
-	usages, ok := fi.snapshot.UsagesByName[symbol]
-	if !ok {
-		return nil, nil
-	}
-	if path == "" {
-		return usages, nil
-	}
-
-	// Filter by search path prefix.
-	var filtered []location
-	for _, loc := range usages {
-		if len(loc.Path) >= len(path) && loc.Path[:len(path)] == path {
-			filtered = append(filtered, loc)
-		}
-	}
-	return filtered, nil
-}
-
-func (fi *fixtureIndexer) IsSymbolUsed(ctx context.Context, name string, hb chan<- struct{}) bool {
-	fi.mu.RLock()
-	defer fi.mu.RUnlock()
-	usages, ok := fi.snapshot.UsagesByName[name]
-	return ok && len(usages) > 0
-}
-
-func (fi *fixtureIndexer) GetImplementations(ctx context.Context, interfaceMethodId string, hb chan<- struct{}) []string {
-	fi.mu.RLock()
-	defer fi.mu.RUnlock()
-	return fi.snapshot.ImplsCache[interfaceMethodId]
-}
-
-func (fi *fixtureIndexer) Packages(ctx context.Context, hb chan<- struct{}) ([]*packages.Package, error) {
-	fi.mu.RLock()
-	defer fi.mu.RUnlock()
-
-	// Build minimal *packages.Package values from the snapshot.
-	// The dead_code pipeline only needs:
-	//   - PkgPath        (for buildFileToPkgMap, isInTargetScope, identifyModule)
-	//   - Module.Path    (for isInTargetScope, identifyModule)
-	//   - GoFiles        (for buildFileToPkgMap, isInTargetScope, hasTextMatchOutsidePackage)
-	//
-	// TypesInfo, Syntax, Fset, Types are left nil. Downstream code
-	// that accesses them (complexity/impact, cross-package AST walk)
-	// has nil guards and gracefully degrades.
-	pkgFiles := make(map[string][]string, len(fi.snapshot.FileToPkg))
-	for file, pkgPath := range fi.snapshot.FileToPkg {
-		pkgFiles[pkgPath] = append(pkgFiles[pkgPath], file)
-	}
-
-	var pkgs []*packages.Package
-	for pkgPath, files := range pkgFiles {
-		pkgs = append(pkgs, &packages.Package{
-			PkgPath: pkgPath,
-			Module: &packages.Module{
-				Path: fi.snapshot.ModulePath,
-			},
-			GoFiles: files,
-		})
-	}
-	return pkgs, nil
-}
-
-func (fi *fixtureIndexer) WarmImplementations(ctx context.Context) {
-	// Fixture impls are pre-computed in the snapshot; no-op.
-}
-
-func (fi *fixtureIndexer) HarvestDeclarations(ctx context.Context, fn func(meta *symMeta) bool, hb chan<- struct{}) error {
-	fi.mu.RLock()
-	defer fi.mu.RUnlock()
-
-	for _, decl := range fi.snapshot.Declarations {
-		if ctx.Err() != nil {
-			return ctx.Err()
-		}
-		if !fn(decl) {
-			return nil
-		}
-	}
-	return nil
-}
