+++ internal/tools/analysis/index_snapshot_test.go
